Ensure refunding bots have enough DYM for gas, some refactoring
